backend/requirements.txt
# Backend requirements for AI Video Analyzer

# MCP (Model Context Protocol)
mcp

# Machine Learning libraries
torch
torchaudio
transformers
accelerate
bitsandbytes

# gRPC transport
grpcio
grpcio-tools
protobuf

# Video processing
moviepy
opencv-python

# Audio processing (additional dependencies for Whisper)
librosa
soundfile

# PDF generation
reportlab  

# Utility libraries
#numpy
#Pillow 

# Development and testing
#pytest
#black
#flake8